Add new select component
