--- src/components/LanguageSelect/LanguageSelect.scss
@@ -1,4 +1,0 @@
-.App {
-  font-family: sans-serif;
-  text-align: center;
-}
